updated delete command

+++ ecommerce_website/management/commands/delete_data.py
@@ -1,54 +1,47 @@
 from django.core.management.base import BaseCommand
 from django.db import connection
-from ecommerce_website.models import Product, ProductSale, ProductAttributeType, ProductAttribute, Order, OrderLine, ProductStock, ProductCategory, ProductFilter
 import os
 from django.conf import settings
+from ecommerce_website.models import (
+    Product, ProductSale, Sale, ProductAttributeType, ProductAttribute,
+    ProductStock, ProductCategory, ProductFilter, Order, OrderLine,
+    DeliveryMethod, StoreMotivation, StoreRating, Brand, ReturnOrderLine, ReturnOrder
+)
 
 
 class Command(BaseCommand):
-    help = 'Delete all data from the database'
+    help = 'Delete all data and seed initial data'
 
     def handle(self, *args, **options):
-
         # Delete existing data
-        Product.objects.all().delete()
-        ProductSale.objects.all().delete()
-        ProductAttributeType.objects.all().delete()
-        ProductAttribute.objects.all().delete()
-        ProductStock.objects.all().delete()
-        ProductCategory.objects.all().delete()
-        ProductFilter.objects.all().delete()
-        Order.objects.all().delete()
-        OrderLine.objects.all().delete()
+        models_to_delete = [
+            Product, ProductSale, Sale, ProductAttributeType, ProductAttribute,
+            ProductStock, ProductCategory, ProductFilter, Order, OrderLine,
+            DeliveryMethod, StoreMotivation, StoreRating, Brand, ReturnOrderLine, ReturnOrder
+        ]
+
+        for model in models_to_delete:
+            model.objects.all().delete()
 
         # Reset primary key sequences for autoincrement fields
         with connection.cursor() as cursor:
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_product';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_productattributetype';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_productattribute';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_productstock';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_productcategory';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_productfilter';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_order';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_orderline';")
-            cursor.execute(
-                "DELETE FROM sqlite_sequence WHERE name='ecommerce_website_productsale';")
+            table_names = [model._meta.db_table for model in models_to_delete]
+            for table_name in table_names:
+                cursor.execute(
+                    f"DELETE FROM sqlite_sequence WHERE name='{table_name}';")
 
-            media_directory = os.path.join(settings.MEDIA_ROOT, 'product_thumbnails')
+        # Delete product thumbnails
+        media_directory = os.path.join(
+            settings.MEDIA_ROOT, 'product_thumbnails')
+        if os.path.isdir(media_directory):
             for filename in os.listdir(media_directory):
                 file_path = os.path.join(media_directory, filename)
                 try:
                     if os.path.isfile(file_path):
                         os.unlink(file_path)
                 except Exception as e:
-                        print(f"Failed to delete {file_path}. Reason: {e}")
+                    self.stderr.write(f"Failed to delete {
+                                      file_path}. Reason: {e}")
 
-            self.stdout.write(self.style.SUCCESS('Seed data and product thumbnails successfully deleted'))
+        self.stdout.write(self.style.SUCCESS(
+            'All data deleted successfully'))
